stopped drawing new circle every repaint

+++ src/main/render/Renderer.kt
@@ -7,7 +7,7 @@ import java.awt.*
 import kotlin.math.roundToInt
 
 object Renderer {
-    fun drawCircle(g: Graphics2D, circle: Circle, screenCenter: Vec2, scale: Double) {
+    fun drawCircle(g: Graphics2D, circle: Circle, screenCenter: Vec2, scale: Float) {
         val cx = (circle.center.x * scale + screenCenter.x).roundToInt()
         val cy = (-circle.center.y * scale + screenCenter.y).roundToInt()
         val r = (circle.radius * scale).roundToInt()
@@ -17,7 +17,7 @@ object Renderer {
         g.drawOval(cx - r, cy - r, r * 2, r * 2)
     }
 
-    fun drawRays(g: Graphics2D, rays: List<Ray>, screenCenter: Vec2, scale: Double, color: Color, strokeWidth: Float = 2f) {
+    fun drawRays(g: Graphics2D, rays: List<Ray>, screenCenter: Vec2, scale: Float, color: Color, strokeWidth: Float = 2f) {
         for (ray in rays) {
             val ox = (ray.origin.x * scale + screenCenter.x).roundToInt()
             val oy = (-ray.origin.y * scale + screenCenter.y).roundToInt()
